Add service handlers

+++ app/models/model_status.go
@@ -12,14 +12,14 @@ package models
 type Status struct {
 
 	// Кол-во пользователей в базе данных.
-	User float32 `json:"user"`
+	User int32 `json:"user"`
 
 	// Кол-во разделов в базе данных.
-	Forum float32 `json:"forum"`
+	Forum int32 `json:"forum"`
 
 	// Кол-во веток обсуждения в базе данных.
-	Thread float32 `json:"thread"`
+	Thread int32 `json:"thread"`
 
 	// Кол-во сообщений в базе данных.
-	Post float32 `json:"post"`
+	Post int64 `json:"post"`
 }
